refactor: incompletedの保管場所を変更
