+++ lib/src/features/home/home_page.dart
@@ -4,7 +4,7 @@ import 'package:todo_app/src/common_widgets/bottom_navigation/bottom_navigation_
 import 'package:todo_app/src/common_widgets/bottom_navigation/bottom_navigation.dart';
 import 'package:todo_app/src/features/todo/presentation/completed/completed_page.dart';
 import 'package:todo_app/src/features/profile/presentation/profile_page.dart';
-import 'package:todo_app/src/features/todo/todo/todo_page.dart';
+import 'package:todo_app/src/features/todo/presentation/incompleted/todo_page.dart';
 
 class HomePage extends ConsumerWidget {
   const HomePage({super.key});
